Consolidate common functions between update and switch platforms
